move clink menu items to single file
