+++ react-app/src/components/Common/topbar.js
@@ -2,14 +2,14 @@ import React from 'react';
 import 'antd/dist/antd.css';
 import '../../index.css';
 import { Layout } from 'antd';
-import EllipsisMenuButton from './ellipsismenubutton';
+import ClinkMenu from './clinkmenu';
 
 const { Header } = Layout;
 
 const Topbar = () => {
   return (
     <Header style={{ position: 'relative', height: '153px' }} >
-      <EllipsisMenuButton />
+      <ClinkMenu />
     </Header>
   )
 }
